refactor(agent): extract FileLock to independent module
Extract FileLock namespace from dispatch.ts to loop/file-lock.ts for better
modularity and code organization. FileLock will be needed by ExecutionSupervisor
when DAG execution logic is moved in subsequent commits.

- Create loop/file-lock.ts with FileLock namespace (acquire, release, owner, clear)
- Update dispatch.ts to import FileLock from new module
- Update loop/index.ts barrel export
- Update test imports in supervisor-worker-split, dispatch, and nested-delegation tests
- No API changes, all 14 tests pass
