+++ packages/agent/src/tools/dispatch.ts
@@ -7,6 +7,7 @@ import type {
   OrchestratorRunInput,
 } from "../loop/orchestration";
 import { RunWorker } from "../loop/run-worker";
+import { FileLock } from "../loop/file-lock";
 import { TaskManager } from "../task/manager";
 import { DispatchInput } from "./schemas";
 import { IngressEngine } from "../ingress/engine";
@@ -109,36 +110,6 @@ interface DispatchOutput {
     error?: string;
   }>;
   error?: string;
-}
-
-export namespace FileLock {
-  const locks = new Map<string, string>();
-
-  export function acquire(filePath: string, agentId: string): boolean {
-    const owner = locks.get(filePath);
-    if (owner && owner !== agentId) {
-      return false;
-    }
-    locks.set(filePath, agentId);
-    return true;
-  }
-
-  export function release(filePath: string, agentId: string): boolean {
-    const owner = locks.get(filePath);
-    if (!owner || owner !== agentId) {
-      return false;
-    }
-    locks.delete(filePath);
-    return true;
-  }
-
-  export function owner(filePath: string): string | undefined {
-    return locks.get(filePath);
-  }
-
-  export function clear(): void {
-    locks.clear();
-  }
 }
 
 export namespace Dispatch {
