+++ packages/agent/test/tools/nested-delegation.test.ts
@@ -4,11 +4,8 @@ import { Session } from "@openomni/session";
 import { BuiltinAgentRegistry } from "../../src/agent/registry";
 import { TaskStorage } from "../../src/task/storage";
 import { Subagent, type SubagentContext } from "../../src/tools/subagent";
-import {
-  Dispatch,
-  FileLock,
-  type DispatchContext,
-} from "../../src/tools/dispatch";
+import { Dispatch, type DispatchContext } from "../../src/tools/dispatch";
+import { FileLock } from "../../src/loop/file-lock";
 
 function createMockLLM() {
   return {
